requirements.txt
supabase
python-dotenv
postgrest